Move class 'UserWithRoleRecord'
Increases cohesion, move record data class to be in the same package
as the DAO that produces the result.
