+++ http-server/src/main/java/org/triplea/server/lobby/chat/ChatParticipantAdapter.java
@@ -4,7 +4,7 @@ import java.util.function.Function;
 import org.triplea.domain.data.PlayerName;
 import org.triplea.http.client.lobby.chat.ChatParticipant;
 import org.triplea.lobby.server.db.data.UserRole;
-import org.triplea.lobby.server.db.data.UserWithRoleRecord;
+import org.triplea.lobby.server.db.dao.api.key.UserWithRoleRecord;
 
 public class ChatParticipantAdapter implements Function<UserWithRoleRecord, ChatParticipant> {
   @Override
