+++ http-server/src/main/java/org/triplea/server/lobby/chat/MessagingService.java
@@ -13,7 +13,7 @@ import org.triplea.http.client.lobby.chat.ChatParticipant;
 import org.triplea.http.client.lobby.chat.events.client.ClientMessageEnvelope;
 import org.triplea.http.client.lobby.chat.events.server.ServerMessageEnvelope;
 import org.triplea.lobby.server.db.dao.api.key.LobbyApiKeyDaoWrapper;
-import org.triplea.lobby.server.db.data.UserWithRoleRecord;
+import org.triplea.lobby.server.db.dao.api.key.UserWithRoleRecord;
 import org.triplea.server.lobby.chat.event.processing.ChatEventProcessor;
 import org.triplea.server.lobby.chat.event.processing.ServerResponse;
