+++ http-server/src/test/java/org/triplea/server/lobby/chat/ChatParticipantAdapterTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.triplea.domain.data.PlayerName;
 import org.triplea.http.client.lobby.chat.ChatParticipant;
 import org.triplea.lobby.server.db.data.UserRole;
-import org.triplea.lobby.server.db.data.UserWithRoleRecord;
+import org.triplea.lobby.server.db.dao.api.key.UserWithRoleRecord;
 
 class ChatParticipantAdapterTest {
