+++ http-server/src/test/java/org/triplea/server/lobby/chat/MessagingServiceTest.java
@@ -23,7 +23,7 @@ import org.triplea.http.client.lobby.chat.events.client.ClientMessageEnvelope;
 import org.triplea.http.client.lobby.chat.events.server.ServerMessageEnvelope;
 import org.triplea.lobby.server.db.dao.api.key.LobbyApiKeyDaoWrapper;
 import org.triplea.lobby.server.db.data.UserRole;
-import org.triplea.lobby.server.db.data.UserWithRoleRecord;
+import org.triplea.lobby.server.db.dao.api.key.UserWithRoleRecord;
 import org.triplea.server.TestData;
 import org.triplea.server.lobby.chat.event.processing.ChatEventProcessor;
 import org.triplea.server.lobby.chat.event.processing.ServerResponse;
