+++ lobby-db-dao/src/main/java/org/triplea/lobby/server/db/JdbiDatabase.java
@@ -14,7 +14,7 @@ import org.triplea.lobby.server.db.data.ModeratorAuditHistoryDaoData;
 import org.triplea.lobby.server.db.data.ModeratorUserDaoData;
 import org.triplea.lobby.server.db.data.UserBanDaoData;
 import org.triplea.lobby.server.db.data.UserRoleLookup;
-import org.triplea.lobby.server.db.data.UserWithRoleRecord;
+import org.triplea.lobby.server.db.dao.api.key.UserWithRoleRecord;
 import org.triplea.lobby.server.db.data.UsernameBanDaoData;
 
 /** Utility to get connections to the Postgres lobby database. */
